Finished all of the refactoring.
