--- Server/methods/validateUser.js
@@ -1,13 +1,0 @@
-'use strict';
-
-const User = require('../api/users/models/User');
-
-module.exports = (user, callback) => {
-  User.count({_id: user}, (err, count) => {
-    if (count > 0) {
-      callback(true);
-    }else {
-      callback(false);
-    }
-  });
-};
